+++ config/nvim/after/ftplugin/rst.vim
@@ -1,0 +1,2 @@
+" Enabled builtin RST indentation support.
+let g:rst_style = 1
