+++ config/nvim/init.lua
@@ -13,7 +13,9 @@ end
 vim.opt.rtp:prepend(lazypath)
 
 -- Load user options and mappings.
-require("user")
+require("user.options")
+require("user.mappings")
+require("user.commands")
 
 -- Load plugins.
 require("lazy").setup("user.plugins")
